moved return type hinting example to a different module, added a py<3.5 compatible equivalent
partly fixes failing tests on py<3.5
